@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
-// Source: vendor/github.com/asecurityteam/runsqs/v3/interface.go
+// Source: /Users/dduncan/go/pkg/mod/github.com/asecurityteam/runsqs/v4@v4.0.0/interface.go
+//
+// Generated by this command:
+//
+//	mockgen -source=/Users/dduncan/go/pkg/mod/github.com/asecurityteam/runsqs/v4@v4.0.0/interface.go -destination=mock_runsqs.go -package=stat
+//
 
 // Package stat is a generated GoMock package.
 package stat
@@ -8,15 +13,17 @@ import (
 	context "context"
 	reflect "reflect"
 
-	runsqs "github.com/asecurityteam/runsqs/v3"
-	sqs "github.com/aws/aws-sdk-go/service/sqs"
-	gomock "github.com/golang/mock/gomock"
+	v4 "github.com/asecurityteam/runsqs/v4"
+	sqs "github.com/aws/aws-sdk-go-v2/service/sqs"
+	types "github.com/aws/aws-sdk-go-v2/service/sqs/types"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockSQSConsumer is a mock of SQSConsumer interface.
 type MockSQSConsumer struct {
 	ctrl     *gomock.Controller
 	recorder *MockSQSConsumerMockRecorder
+	isgomock struct{}
 }
 
 // MockSQSConsumerMockRecorder is the mock recorder for MockSQSConsumer.
@@ -37,10 +44,10 @@ func (m *MockSQSConsumer) EXPECT() *MockSQSConsumerMockRecorder {
 }
 
 // GetSQSMessageConsumer mocks base method.
-func (m *MockSQSConsumer) GetSQSMessageConsumer() runsqs.SQSMessageConsumer {
+func (m *MockSQSConsumer) GetSQSMessageConsumer() v4.SQSMessageConsumer {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetSQSMessageConsumer")
-	ret0, _ := ret[0].(runsqs.SQSMessageConsumer)
+	ret0, _ := ret[0].(v4.SQSMessageConsumer)
 	return ret0
 }
 
@@ -59,7 +66,7 @@ func (m *MockSQSConsumer) StartConsuming(ctx context.Context) error {
 }
 
 // StartConsuming indicates an expected call of StartConsuming.
-func (mr *MockSQSConsumerMockRecorder) StartConsuming(ctx interface{}) *gomock.Call {
+func (mr *MockSQSConsumerMockRecorder) StartConsuming(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "StartConsuming", reflect.TypeOf((*MockSQSConsumer)(nil).StartConsuming), ctx)
 }
@@ -73,7 +80,7 @@ func (m *MockSQSConsumer) StopConsuming(ctx context.Context) error {
 }
 
 // StopConsuming indicates an expected call of StopConsuming.
-func (mr *MockSQSConsumerMockRecorder) StopConsuming(ctx interface{}) *gomock.Call {
+func (mr *MockSQSConsumerMockRecorder) StopConsuming(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "StopConsuming", reflect.TypeOf((*MockSQSConsumer)(nil).StopConsuming), ctx)
 }
@@ -82,6 +89,7 @@ func (mr *MockSQSConsumerMockRecorder) StopConsuming(ctx interface{}) *gomock.Ca
 type MockSQSMessageConsumer struct {
 	ctrl     *gomock.Controller
 	recorder *MockSQSMessageConsumerMockRecorder
+	isgomock struct{}
 }
 
 // MockSQSMessageConsumerMockRecorder is the mock recorder for MockSQSMessageConsumer.
@@ -102,27 +110,27 @@ func (m *MockSQSMessageConsumer) EXPECT() *MockSQSMessageConsumerMockRecorder {
 }
 
 // ConsumeMessage mocks base method.
-func (m *MockSQSMessageConsumer) ConsumeMessage(ctx context.Context, message *sqs.Message) runsqs.SQSMessageConsumerError {
+func (m *MockSQSMessageConsumer) ConsumeMessage(ctx context.Context, message *types.Message) v4.SQSMessageConsumerError {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ConsumeMessage", ctx, message)
-	ret0, _ := ret[0].(runsqs.SQSMessageConsumerError)
+	ret0, _ := ret[0].(v4.SQSMessageConsumerError)
 	return ret0
 }
 
 // ConsumeMessage indicates an expected call of ConsumeMessage.
-func (mr *MockSQSMessageConsumerMockRecorder) ConsumeMessage(ctx, message interface{}) *gomock.Call {
+func (mr *MockSQSMessageConsumerMockRecorder) ConsumeMessage(ctx, message any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ConsumeMessage", reflect.TypeOf((*MockSQSMessageConsumer)(nil).ConsumeMessage), ctx, message)
 }
 
 // DeadLetter mocks base method.
-func (m *MockSQSMessageConsumer) DeadLetter(ctx context.Context, message *sqs.Message) {
+func (m *MockSQSMessageConsumer) DeadLetter(ctx context.Context, message *types.Message) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "DeadLetter", ctx, message)
 }
 
 // DeadLetter indicates an expected call of DeadLetter.
-func (mr *MockSQSMessageConsumerMockRecorder) DeadLetter(ctx, message interface{}) *gomock.Call {
+func (mr *MockSQSMessageConsumerMockRecorder) DeadLetter(ctx, message any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeadLetter", reflect.TypeOf((*MockSQSMessageConsumer)(nil).DeadLetter), ctx, message)
 }
@@ -131,6 +139,7 @@ func (mr *MockSQSMessageConsumerMockRecorder) DeadLetter(ctx, message interface{
 type MockSQSMessageConsumerError struct {
 	ctrl     *gomock.Controller
 	recorder *MockSQSMessageConsumerErrorMockRecorder
+	isgomock struct{}
 }
 
 // MockSQSMessageConsumerErrorMockRecorder is the mock recorder for MockSQSMessageConsumerError.
@@ -179,10 +188,10 @@ func (mr *MockSQSMessageConsumerErrorMockRecorder) IsRetryable() *gomock.Call {
 }
 
 // RetryAfter mocks base method.
-func (m *MockSQSMessageConsumerError) RetryAfter() int64 {
+func (m *MockSQSMessageConsumerError) RetryAfter() int32 {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "RetryAfter")
-	ret0, _ := ret[0].(int64)
+	ret0, _ := ret[0].(int32)
 	return ret0
 }
 
@@ -196,6 +205,7 @@ func (mr *MockSQSMessageConsumerErrorMockRecorder) RetryAfter() *gomock.Call {
 type MockSQSProducer struct {
 	ctrl     *gomock.Controller
 	recorder *MockSQSProducerMockRecorder
+	isgomock struct{}
 }
 
 // MockSQSProducerMockRecorder is the mock recorder for MockSQSProducer.
@@ -215,6 +225,21 @@ func (m *MockSQSProducer) EXPECT() *MockSQSProducerMockRecorder {
 	return m.recorder
 }
 
+// BatchProduceMessage mocks base method.
+func (m *MockSQSProducer) BatchProduceMessage(ctx context.Context, messageBatchInput *sqs.SendMessageBatchInput) (*sqs.SendMessageBatchOutput, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "BatchProduceMessage", ctx, messageBatchInput)
+	ret0, _ := ret[0].(*sqs.SendMessageBatchOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// BatchProduceMessage indicates an expected call of BatchProduceMessage.
+func (mr *MockSQSProducerMockRecorder) BatchProduceMessage(ctx, messageBatchInput any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "BatchProduceMessage", reflect.TypeOf((*MockSQSProducer)(nil).BatchProduceMessage), ctx, messageBatchInput)
+}
+
 // ProduceMessage mocks base method.
 func (m *MockSQSProducer) ProduceMessage(ctx context.Context, messageInput *sqs.SendMessageInput) error {
 	m.ctrl.T.Helper()
@@ -224,7 +249,7 @@ func (m *MockSQSProducer) ProduceMessage(ctx context.Context, messageInput *sqs.
 }
 
 // ProduceMessage indicates an expected call of ProduceMessage.
-func (mr *MockSQSProducerMockRecorder) ProduceMessage(ctx, messageInput interface{}) *gomock.Call {
+func (mr *MockSQSProducerMockRecorder) ProduceMessage(ctx, messageInput any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ProduceMessage", reflect.TypeOf((*MockSQSProducer)(nil).ProduceMessage), ctx, messageInput)
 }
@@ -241,4 +266,128 @@ func (m *MockSQSProducer) QueueURL() string {
 func (mr *MockSQSProducerMockRecorder) QueueURL() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueueURL", reflect.TypeOf((*MockSQSProducer)(nil).QueueURL))
+}
+
+// MockSQSClient is a mock of SQSClient interface.
+type MockSQSClient struct {
+	ctrl     *gomock.Controller
+	recorder *MockSQSClientMockRecorder
+	isgomock struct{}
+}
+
+// MockSQSClientMockRecorder is the mock recorder for MockSQSClient.
+type MockSQSClientMockRecorder struct {
+	mock *MockSQSClient
+}
+
+// NewMockSQSClient creates a new mock instance.
+func NewMockSQSClient(ctrl *gomock.Controller) *MockSQSClient {
+	mock := &MockSQSClient{ctrl: ctrl}
+	mock.recorder = &MockSQSClientMockRecorder{mock}
+	return mock
+}
+
+// EXPECT returns an object that allows the caller to indicate expected use.
+func (m *MockSQSClient) EXPECT() *MockSQSClientMockRecorder {
+	return m.recorder
+}
+
+// ChangeMessageVisibility mocks base method.
+func (m *MockSQSClient) ChangeMessageVisibility(ctx context.Context, params *sqs.ChangeMessageVisibilityInput, optFns ...func(*sqs.Options)) (*sqs.ChangeMessageVisibilityOutput, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, params}
+	for _, a := range optFns {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ChangeMessageVisibility", varargs...)
+	ret0, _ := ret[0].(*sqs.ChangeMessageVisibilityOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ChangeMessageVisibility indicates an expected call of ChangeMessageVisibility.
+func (mr *MockSQSClientMockRecorder) ChangeMessageVisibility(ctx, params any, optFns ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, params}, optFns...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ChangeMessageVisibility", reflect.TypeOf((*MockSQSClient)(nil).ChangeMessageVisibility), varargs...)
+}
+
+// DeleteMessage mocks base method.
+func (m *MockSQSClient) DeleteMessage(ctx context.Context, params *sqs.DeleteMessageInput, optFns ...func(*sqs.Options)) (*sqs.DeleteMessageOutput, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, params}
+	for _, a := range optFns {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "DeleteMessage", varargs...)
+	ret0, _ := ret[0].(*sqs.DeleteMessageOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// DeleteMessage indicates an expected call of DeleteMessage.
+func (mr *MockSQSClientMockRecorder) DeleteMessage(ctx, params any, optFns ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, params}, optFns...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteMessage", reflect.TypeOf((*MockSQSClient)(nil).DeleteMessage), varargs...)
+}
+
+// ReceiveMessage mocks base method.
+func (m *MockSQSClient) ReceiveMessage(ctx context.Context, params *sqs.ReceiveMessageInput, optFns ...func(*sqs.Options)) (*sqs.ReceiveMessageOutput, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, params}
+	for _, a := range optFns {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ReceiveMessage", varargs...)
+	ret0, _ := ret[0].(*sqs.ReceiveMessageOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ReceiveMessage indicates an expected call of ReceiveMessage.
+func (mr *MockSQSClientMockRecorder) ReceiveMessage(ctx, params any, optFns ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, params}, optFns...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReceiveMessage", reflect.TypeOf((*MockSQSClient)(nil).ReceiveMessage), varargs...)
+}
+
+// SendMessage mocks base method.
+func (m *MockSQSClient) SendMessage(ctx context.Context, params *sqs.SendMessageInput, optFns ...func(*sqs.Options)) (*sqs.SendMessageOutput, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, params}
+	for _, a := range optFns {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "SendMessage", varargs...)
+	ret0, _ := ret[0].(*sqs.SendMessageOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// SendMessage indicates an expected call of SendMessage.
+func (mr *MockSQSClientMockRecorder) SendMessage(ctx, params any, optFns ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, params}, optFns...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendMessage", reflect.TypeOf((*MockSQSClient)(nil).SendMessage), varargs...)
+}
+
+// SendMessageBatch mocks base method.
+func (m *MockSQSClient) SendMessageBatch(ctx context.Context, params *sqs.SendMessageBatchInput, optFns ...func(*sqs.Options)) (*sqs.SendMessageBatchOutput, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, params}
+	for _, a := range optFns {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "SendMessageBatch", varargs...)
+	ret0, _ := ret[0].(*sqs.SendMessageBatchOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// SendMessageBatch indicates an expected call of SendMessageBatch.
+func (mr *MockSQSClientMockRecorder) SendMessageBatch(ctx, params any, optFns ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, params}, optFns...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendMessageBatch", reflect.TypeOf((*MockSQSClient)(nil).SendMessageBatch), varargs...)
 }
